Fix issue #438 - OneDev upgrade does not work if server is terminated forcibly

+++ server-core/src/main/java/io/onedev/server/maintenance/Upgrade.java
@@ -179,12 +179,8 @@ public class Upgrade extends DefaultPersistManager {
 			if (oldAppVersion == null || !oldAppVersion.equals(AppLoader.getProduct().getVersion())) {
 				logger.info("Upgrading {}...", upgradeDir.getAbsolutePath());
 				
-				if (Bootstrap.isServerRunning(upgradeDir)) {
-					logger.error("Please stop server running at \"{}\" before upgrading", upgradeDir.getAbsolutePath());
-					System.exit(1);
-				}
-				if (Bootstrap.isServerRunning(Bootstrap.installDir)) {
-					logger.error("Please stop server running at \"{}\" before upgrading", Bootstrap.installDir.getAbsolutePath());
+				if (isServerRunning(upgradeDir) || isServerRunning(Bootstrap.installDir)) {
+					logger.error("Please stop server before upgrading");
 					System.exit(1);
 				}
 
@@ -395,13 +391,14 @@ public class Upgrade extends DefaultPersistManager {
 		} else {
 			logger.info("Populating {}...", upgradeDir.getAbsolutePath());
 			
-			if (Bootstrap.isServerRunning(Bootstrap.installDir)) {
-				logger.error("Please stop server running at \"{}\" before populating", Bootstrap.installDir.getAbsolutePath());
+			if (isServerRunning(Bootstrap.installDir)) {
+				logger.error("Please stop server before populating");
 				System.exit(1);
 			}
 
 			try {
 				FileUtils.copyDirectory(Bootstrap.installDir, upgradeDir);
+				FileUtils.cleanDir(new File(upgradeDir, "logs"));
 				restoreExecutables(upgradeDir);
 				logger.info("Successfully populated {}", upgradeDir.getAbsolutePath());
 				System.exit(0);
@@ -418,36 +415,38 @@ public class Upgrade extends DefaultPersistManager {
 		
 		File siteServerScriptFile = new File(upgradeDir, "bin/server.sh");
 		for (File file: Bootstrap.getBinDir().listFiles()) {
-			try {
-				String serverScript = null;
-				if (file.getName().equals("server.sh")) {
-					String siteRunAsUserLine = null;
-					for (String line: (List<String>)FileUtils.readLines(siteServerScriptFile, Charset.defaultCharset())) {
-						if (line.contains("RUN_AS_USER")) {
-							siteRunAsUserLine = line;
-							break;
-						}
-					}
-					if (siteRunAsUserLine != null) {
-						String newRunAsUserLine = null;
-						for (String line: (List<String>)FileUtils.readLines(file, Charset.defaultCharset())) {
+			if (!file.getName().endsWith(".pid") && !file.getName().endsWith(".status")) {
+				try {
+					String serverScript = null;
+					if (file.getName().equals("server.sh")) {
+						String siteRunAsUserLine = null;
+						for (String line: (List<String>)FileUtils.readLines(siteServerScriptFile, Charset.defaultCharset())) {
 							if (line.contains("RUN_AS_USER")) {
-								newRunAsUserLine = line;
+								siteRunAsUserLine = line;
 								break;
 							}
 						}
-						if (newRunAsUserLine != null) {
-							serverScript = FileUtils.readFileToString(file, Charset.defaultCharset());
-							serverScript = StringUtils.replace(serverScript, newRunAsUserLine, siteRunAsUserLine);
+						if (siteRunAsUserLine != null) {
+							String newRunAsUserLine = null;
+							for (String line: (List<String>)FileUtils.readLines(file, Charset.defaultCharset())) {
+								if (line.contains("RUN_AS_USER")) {
+									newRunAsUserLine = line;
+									break;
+								}
+							}
+							if (newRunAsUserLine != null) {
+								serverScript = FileUtils.readFileToString(file, Charset.defaultCharset());
+								serverScript = StringUtils.replace(serverScript, newRunAsUserLine, siteRunAsUserLine);
+							}
 						}
 					}
+					if (serverScript != null)
+						FileUtils.writeStringToFile(siteServerScriptFile, serverScript, Charset.defaultCharset());
+					else 
+						FileUtils.copyFile(file, new File(upgradeDir, "bin/" + file.getName()));
+				} catch (IOException e) {
+					throw new RuntimeException(e);
 				}
-				if (serverScript != null)
-					FileUtils.writeStringToFile(siteServerScriptFile, serverScript, Charset.defaultCharset());
-				else 
-					FileUtils.copyFile(file, new File(upgradeDir, "bin/" + file.getName()));
-			} catch (IOException e) {
-				throw new RuntimeException(e);
 			}
 		}
 
@@ -468,16 +467,20 @@ public class Upgrade extends DefaultPersistManager {
 		if (new File(upgradeDir, "bin/reset_admin_password.sh").exists())
 			FileUtils.deleteFile(new File(upgradeDir, "bin/reset_admin_password.sh"));
 		
+		FileUtils.deleteDir(new File(upgradeDir, "status"));
+		
 		cleanAndCopy(Bootstrap.getBootDir(), new File(upgradeDir, "boot"));
 		if (new File(upgradeDir, "boot/system.classpath").exists())
 			FileUtils.deleteFile(new File(upgradeDir, "boot/system.classpath"));
 		
 		cleanAndCopy(Bootstrap.getLibDir(), new File(upgradeDir, "lib"));
 		for (File file: new File(Bootstrap.getSiteDir(), "avatars").listFiles()) {
-			try {
-				FileUtils.copyFileToDirectory(file, new File(upgradeDir, "site/avatars"));
-			} catch (IOException e) {
-				throw new RuntimeException(e);
+			if (file.isFile()) {
+				try {
+					FileUtils.copyFileToDirectory(file, new File(upgradeDir, "site/avatars"));
+				} catch (IOException e) {
+					throw new RuntimeException(e);
+				}
 			}
 		}
 		File uploadedDir = new File(upgradeDir, "site/avatars/uploaded");
@@ -507,6 +510,8 @@ public class Upgrade extends DefaultPersistManager {
 					"io.onedev.commons.launcher.bootstrap.Bootstrap");
 			wrapperConf = StringUtils.replace(wrapperConf, "io.onedev.commons.launcher.bootstrap.Bootstrap", 
 					"io.onedev.commons.bootstrap.Bootstrap");
+			wrapperConf = StringUtils.replace(wrapperConf, "wrapper.pidfile=../status/onedev.pid", "");
+			
 			FileUtils.writeStringToFile(wrapperConfFile, wrapperConf, StandardCharsets.UTF_8);
 			
 			File hibernatePropsFile = new File(upgradeDir, "conf/hibernate.properties");
